requirements.txt
biopython
condacolab
jupyter
kaleido
matplotlib
pandas
paperclip
plotly
progressbar
progressbar2
pyautogui
scikit-learn